requirements.txt
# Core Dependencies
ddgs>=7.5.0,<8.0.0
python-dateutil>=2.8.0,<3.0.0
trafilatura>=1.12.0,<2.0.0
deep-translator>=1.11.0,<2.0.0
requests>=2.31.0,<3.0.0
PyYAML>=6.0.0,<7.0.0

# AI & Analysis
groq>=0.11.0,<1.0.0
google-generativeai>=0.8.0,<1.0.0
textblob>=0.18.0,<1.0.0

# Rich Console UI
rich>=13.0.0,<14.0.0

# Environment Variables
python-dotenv>=1.0.0,<2.0.0
pyperclip>=1.8.0,<2.0.0
